requirements.txt
python-dotenv>=1.0.0
SQLAlchemy>=2.0.25
psycopg2-binary>=2.9.9
pandas>=2.2.0
ipython-sql>=0.5.0
jupyter>=1.0.0
